- Changed all Point objects to Point2f in the ClusterDetector module - Added an extra validation step in the *MatFactory classes - Added a validation step to the method for saving output image to disk in the SimulationClusterDetector class

+++ Multiscale/include/multiscale/analysis/spatial/cluster/SimulationClusterDetector.hpp
@@ -27,14 +27,14 @@ namespace multiscale {
 
             private:
 
-                Mat thresholdedImage;           /*!< Thresholded version of the image */
-                Mat outputImage;                /*!< Image for displaying the results */
+                Mat thresholdedImage;     /*!< Thresholded version of the image */
+                Mat outputImage;          /*!< Image for displaying the results */
 
-                unsigned int height;            /*!< Height of the grid used in the simulation */
-                unsigned int width;             /*!< Width of the grid used in the simulation */
+                unsigned int height;      /*!< Height of the grid used in the simulation */
+                unsigned int width;       /*!< Width of the grid used in the simulation */
 
-                unsigned int entityHeight;      /*!< Height of an entity */
-                unsigned int entityWidth;       /*!< Width of an entity */
+                double entityHeight;      /*!< Height of an entity */
+                double entityWidth;       /*!< Width of an entity */
 
             public:
